Class PGSSPDE is now included in SPDE
